sessionコントローラーを無くし、内容をアプリケーションコントローラーへ移植、それに伴いビュー、ヘルパー、テストを削除
